pyproject.toml
[project]
name = "hit-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.1.0",
    "tiktoken>=0.11.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-milvus>=0.2.0",
    "langchain-core>=0.3.0",
    "pymilvus>=2.4.0",
    "numpy>=1.24.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
]
